Remove java-version-checker tool and move the version check into the respective initialization scripts
