+++ distribution/build.gradle
@@ -261,7 +261,6 @@ configure(subprojects.findAll { ['archives', 'packages'].contains(it.name) }) {
         from { project(':server').jar }
         from { project(':server').configurations.runtime }
         from { project(':libs:elasticsearch-plugin-classloader').jar }
-        from { project(':distribution:tools:java-version-checker').jar }
         from { project(':distribution:tools:launchers').jar }
         into('tools/plugin-cli') {
           from { project(':distribution:tools:plugin-cli').jar }
