--- distribution/tools/java-version-checker/build.gradle
@@ -1,13 +1,0 @@
-apply plugin: 'elasticsearch.build'
-
-targetCompatibility = JavaVersion.VERSION_1_7
-
-// java_version_checker do not depend on core so only JDK signatures should be checked
-forbiddenApisMain {
-  replaceSignatureFiles 'jdk-signatures'
-}
-
-test.enabled = false
-javadoc.enabled = false
-loggerUsageCheck.enabled = false
-jarHell.enabled = false
