+++ distribution/tools/launchers/src/main/java/org/elasticsearch/tools/launchers/TempDirectory.java
@@ -20,10 +20,17 @@
 package org.elasticsearch.tools.launchers;
 
 import java.io.IOException;
-import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Arrays;
+import java.util.Locale;
+
+import static java.lang.System.err;
+import static java.lang.System.exit;
+import static java.lang.System.getProperty;
+import static java.lang.System.out;
+import static java.nio.file.Files.createDirectories;
+import static java.nio.file.Files.createTempDirectory;
 
 /**
  * Provides a path for a temporary directory. On non-Windows OS, this will be created as a sub-directory of the default temporary directory.
@@ -38,22 +45,28 @@ final class TempDirectory {
      * @param args the args to the program which should be empty
      * @throws IOException if an I/O exception occurred while creating the temporary directory
      */
+    @SuppressForbidden(reason = "System#err, System#exit, System#out")
     public static void main(final String[] args) throws IOException {
         if (args.length != 0) {
-            throw new IllegalArgumentException("expected zero arguments but was " + Arrays.toString(args));
+            err.println(String.format(
+                Locale.ROOT,
+                "expected zero arguments but was %s",
+                Arrays.toString(args)
+
+            ));
+            exit(1);
         }
+
         /*
          * On Windows, we avoid creating a unique temporary directory per invocation lest we pollute the temporary directory. On other
          * operating systems, temporary directories will be cleaned automatically via various mechanisms (e.g., systemd, or restarts).
          */
-        final Path path;
-        if (System.getProperty("os.name").startsWith("Windows")) {
-            path = Paths.get(System.getProperty("java.io.tmpdir"), "elasticsearch");
-            Files.createDirectories(path);
-        } else {
-            path = Launchers.createTempDirectory("elasticsearch-");
-        }
-        Launchers.outPrintln(path.toString());
+        final Path path = getProperty("os.name").startsWith("Windows")?
+            createDirectories(Paths.get(getProperty("java.io.tmpdir"), "elasticsearch")):
+            createTempDirectory("elasticsearch-");
+
+        out.println(path.toString());
+        exit(0);
     }
 
 }
